refactor(settings): move org settings to /org/settings/[tab] with URL-driven routing
Replace client-side tab switching with server-side URL routing for org
settings. Each tab now gets its own URL and only fetches the data it
needs. Old /settings routes redirect to the new paths for backward compat.
