+++ app/(app)/org/settings/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function OrgSettingsPage() {
+  redirect("/org/settings/variables");
+}
